move component files to individual folders. rename Agent class to AgentFactory.
